pyproject.toml
[project]
name = "student-schedule-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "dj-database-url>=3.0.0",
    "django-ninja>=1.4.3",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "python-telegram-bot>=22.1",
]


[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
extend-select = [
  "UP",
  "E",
  "W",
  "I",
  "C90",
  "N",
  "ANN",
  "ASYNC",
  "BLE",
  "FBT",
  "B",
  "A",
  "C4",
  "DTZ",
  "T10",
  "FA",
  "ISC",
  "ICN",
  "INP",
  "PIE",
  "T20",
  "PYI",
  "Q",
  "RET",
  "SLF",
  "SIM",
  "TID",
  "TCH",
  "ARG",
  "TD",
  "FIX",
  "ERA",
  "PL",
  "RUF",
]
ignore = [
  "ISC001",
  "N999",
  "ERA001",
  "RUF012",
  "ANN101",
  "ANN102",
  "FBT001",
  "FBT002",
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true